+++ main.ts
@@ -6,6 +6,7 @@ interface OpenWeatherSettings {
   location: string;
   key: string;
   units: string;
+  language: string;
   excludeFolder: string;
   weatherFormat1: string;
   weatherFormat2: string;
@@ -20,6 +21,7 @@ const DEFAULT_SETTINGS: OpenWeatherSettings = {
   location: '',
   key: '',
   units: 'metric',
+  language: 'en',
   excludeFolder: '',
   weatherFormat1: '%desc% • Current Temp: %temp%°C • Feels Like: %feels%°C\n',
   weatherFormat2: '%name%: %dateMonth4% %dateDay2% - %timeH2%:%timeM% %ampm1%\nCurrent Temp: %temp%°C • Feels Like: %feels%°C\nWind: %wind-speed% km/h from the %wind-dir%^ with gusts up to %wind-gust% km/h^\nSunrise: %sunrise% • Sunset: %sunset%\n',
@@ -39,12 +41,14 @@ class FormatWeather {
   location: String;
   key: string
   units: string
+  language: string
   format: string
   
-  constructor(location: string, key: string, units: string, format: string) {
+  constructor(location: string, key: string, units: string, language: string, format: string) {
     this.location = location;
     this.key = key;
     this.units = units;
+    this.language = language;
     this.format = format;
   }
   
@@ -52,7 +56,7 @@ class FormatWeather {
   async getWeather() {
     let weatherData;
     let weatherString;
-    let url = `https://api.openweathermap.org/data/2.5/weather?q=${this.location}&appid=${this.key}&units=${this.units}`;
+    let url = `https://api.openweathermap.org/data/2.5/weather?q=${this.location}&lang=${this.language}&appid=${this.key}&units=${this.units}`;
     let req = await fetch(url);
     let json = await req.json();
     let conditions = json.weather[0].description;
@@ -336,7 +340,7 @@ export default class OpenWeather extends Plugin {
         new Notice("Open a Markdown file first.");
         return;
       }
-      new InsertWeatherPicker(this.settings.location, this.settings.key, this.settings.units, this.settings.weatherFormat1, this.settings.weatherFormat2, this.settings.weatherFormat3, this.settings.weatherFormat4).open();
+      new InsertWeatherPicker(this.settings.location, this.settings.key, this.settings.units, this.settings.language, this.settings.weatherFormat1, this.settings.weatherFormat2, this.settings.weatherFormat3, this.settings.weatherFormat4).open();
     });
 
     // onload - This adds a status bar item to the bottom of the app - Does not work on mobile apps 
@@ -350,7 +354,7 @@ export default class OpenWeather extends Plugin {
           displayErrorMsg = false;
         }
       } else {
-        let wstr = new FormatWeather(this.settings.location, this.settings.key, this.settings.units, this.settings.weatherFormatSB);
+        let wstr = new FormatWeather(this.settings.location, this.settings.key, this.settings.units, this.settings.language, this.settings.weatherFormatSB);
         let weatherStr = await wstr.getWeatherString();
         if (weatherStr.length == 0) {return};
         this.statusBar.setText(weatherStr);
@@ -366,7 +370,7 @@ export default class OpenWeather extends Plugin {
       editorCallback: async (editor: Editor, view: MarkdownView) => {
         if (this.settings.weatherFormat1.length > 0) {
           if (view.data.contains("%weather1%")) {
-            let wstr = new FormatWeather (this.settings.location, this.settings.key, this.settings.units, this.settings.weatherFormat1);
+            let wstr = new FormatWeather (this.settings.location, this.settings.key, this.settings.units, this.settings.language, this.settings.weatherFormat1);
             let weatherStr = await wstr.getWeatherString();
             if (weatherStr.length == 0) {return};
             let doc = editor.getValue().replace(/%weather1%/gmi, weatherStr);
@@ -375,7 +379,7 @@ export default class OpenWeather extends Plugin {
         }
         if (this.settings.weatherFormat2.length > 0) {
           if (view.data.contains("%weather2%")) {
-            let wstr = new FormatWeather (this.settings.location, this.settings.key, this.settings.units, this.settings.weatherFormat2);
+            let wstr = new FormatWeather (this.settings.location, this.settings.key, this.settings.units, this.settings.language, this.settings.weatherFormat2);
             let weatherStr = await wstr.getWeatherString();
             if (weatherStr.length == 0) {return};
             let doc = editor.getValue().replace(/%weather2%/gmi, weatherStr);
@@ -384,7 +388,7 @@ export default class OpenWeather extends Plugin {
         }
         if (this.settings.weatherFormat3.length > 0) {
           if (view.data.contains("%weather3%")) {
-            let wstr = new FormatWeather (this.settings.location, this.settings.key, this.settings.units, this.settings.weatherFormat3);
+            let wstr = new FormatWeather (this.settings.location, this.settings.key, this.settings.units, this.settings.language, this.settings.weatherFormat3);
             let weatherStr = await wstr.getWeatherString();
             if (weatherStr.length == 0) {return};
             let doc = editor.getValue().replace(/%weather3%/gmi, weatherStr);
@@ -393,7 +397,7 @@ export default class OpenWeather extends Plugin {
         }
         if (this.settings.weatherFormat4.length > 0) {
           if (view.data.contains("%weather4%")) {
-            let wstr = new FormatWeather (this.settings.location, this.settings.key, this.settings.units, this.settings.weatherFormat4);
+            let wstr = new FormatWeather (this.settings.location, this.settings.key, this.settings.units, this.settings.language, this.settings.weatherFormat4);
             let weatherStr = await wstr.getWeatherString();
             if (weatherStr.length == 0) {return};
             let doc = editor.getValue().replace(/%weather4%/gmi, weatherStr);
@@ -409,7 +413,7 @@ export default class OpenWeather extends Plugin {
       name: 'Insert weather format one',
       editorCallback: async (editor: Editor, view: MarkdownView) => {
         if (this.settings.weatherFormat1.length > 0) {
-          let wstr = new FormatWeather (this.settings.location, this.settings.key, this.settings.units, this.settings.weatherFormat1);
+          let wstr = new FormatWeather (this.settings.location, this.settings.key, this.settings.units, this.settings.language, this.settings.weatherFormat1);
           let weatherStr = await wstr.getWeatherString();
           if (weatherStr.length == 0) {return};
           editor.replaceSelection(`${weatherStr}`);
@@ -425,7 +429,7 @@ export default class OpenWeather extends Plugin {
       name: 'Insert weather format two',
       editorCallback: async (editor: Editor, view: MarkdownView) => {
         if (this.settings.weatherFormat2.length > 0) {
-          let wstr = new FormatWeather (this.settings.location, this.settings.key, this.settings.units, this.settings.weatherFormat2);
+          let wstr = new FormatWeather (this.settings.location, this.settings.key, this.settings.units, this.settings.language, this.settings.weatherFormat2);
           let weatherStr = await wstr.getWeatherString();
           if (weatherStr.length == 0) {return};
           editor.replaceSelection(`${weatherStr}`);
@@ -441,7 +445,7 @@ export default class OpenWeather extends Plugin {
       name: 'Insert weather format three',
       editorCallback: async (editor: Editor, view: MarkdownView) => {
         if (this.settings.weatherFormat3.length > 0) {
-          let wstr = new FormatWeather (this.settings.location, this.settings.key, this.settings.units, this.settings.weatherFormat3);
+          let wstr = new FormatWeather (this.settings.location, this.settings.key, this.settings.units, this.settings.language, this.settings.weatherFormat3);
           let weatherStr = await wstr.getWeatherString();
           if (weatherStr.length == 0) {return};
           editor.replaceSelection(`${weatherStr}`);
@@ -457,7 +461,7 @@ export default class OpenWeather extends Plugin {
       name: 'Insert weather format four',
       editorCallback: async (editor: Editor, view: MarkdownView) => {
         if (this.settings.weatherFormat4.length > 0) {
-          let wstr = new FormatWeather (this.settings.location, this.settings.key, this.settings.units, this.settings.weatherFormat4);
+          let wstr = new FormatWeather (this.settings.location, this.settings.key, this.settings.units, this.settings.language, this.settings.weatherFormat4);
           let weatherStr = await wstr.getWeatherString();
           if (weatherStr.length == 0) {return};
           editor.replaceSelection(`${weatherStr}`);
@@ -505,27 +509,27 @@ export default class OpenWeather extends Plugin {
     if (!view) return;
       if(document.getElementsByClassName('weather_current_1').length === 1) {
         const divEl = document.getElementsByClassName('weather_current_1')[0];
-        let wstr = new FormatWeather (this.settings.location, this.settings.key, this.settings.units, this.settings.weatherFormat1);
+        let wstr = new FormatWeather (this.settings.location, this.settings.key, this.settings.units, this.settings.language, this.settings.weatherFormat1);
         let weatherStr = await wstr.getWeatherString();
         if (weatherStr.length == 0) {return};
         divEl.innerHTML = weatherStr;
       }
       if(document.getElementsByClassName('weather_current_2').length === 1) {
         const divEl = document.getElementsByClassName('weather_current_2')[0];
-        let wstr = new FormatWeather (this.settings.location, this.settings.key, this.settings.units, this.settings.weatherFormat2);
+        let wstr = new FormatWeather (this.settings.location, this.settings.key, this.settings.units, this.settings.language, this.settings.weatherFormat2);
         let weatherStr = await wstr.getWeatherString();
         if (weatherStr.length == 0) {return};
         divEl.innerHTML = weatherStr;
       }
       if(document.getElementsByClassName('weather_current_3').length === 1) {
         const divEl = document.getElementsByClassName('weather_current_3')[0];
-        let wstr = new FormatWeather (this.settings.location, this.settings.key, this.settings.units, this.settings.weatherFormat3);
+        let wstr = new FormatWeather (this.settings.location, this.settings.key, this.settings.units, this.settings.language, this.settings.weatherFormat3);
         let weatherStr = await wstr.getWeatherString();
         divEl.innerHTML = weatherStr;
       }
       if(document.getElementsByClassName('weather_current_4').length === 1) {
         const divEl = document.getElementsByClassName('weather_current_4')[0];
-        let wstr = new FormatWeather (this.settings.location, this.settings.key, this.settings.units, this.settings.weatherFormat4);
+        let wstr = new FormatWeather (this.settings.location, this.settings.key, this.settings.units, this.settings.language, this.settings.weatherFormat4);
         let weatherStr = await wstr.getWeatherString();
         if (weatherStr.length == 0) {return};
         divEl.innerHTML = weatherStr;
@@ -542,7 +546,7 @@ export default class OpenWeather extends Plugin {
     if (editor == null) return;
     if (this.settings.weatherFormat1.length > 0) {
       if (editor.contains("%weather1%")) {
-        let wstr = new FormatWeather (this.settings.location, this.settings.key, this.settings.units, this.settings.weatherFormat1);
+        let wstr = new FormatWeather (this.settings.location, this.settings.key, this.settings.units, this.settings.language, this.settings.weatherFormat1);
         let weatherStr = await wstr.getWeatherString();
         if (weatherStr.length == 0) {return};
         editor = editor.replace(/%weather1%/gmi, weatherStr);
@@ -551,7 +555,7 @@ export default class OpenWeather extends Plugin {
     }
     if (this.settings.weatherFormat2.length > 0) {
       if (editor.contains("%weather2%")) {
-        let wstr = new FormatWeather (this.settings.location, this.settings.key, this.settings.units, this.settings.weatherFormat2);
+        let wstr = new FormatWeather (this.settings.location, this.settings.key, this.settings.units, this.settings.language, this.settings.weatherFormat2);
         let weatherStr = await wstr.getWeatherString();
         if (weatherStr.length == 0) {return};
         editor = editor.replace(/%weather2%/gmi, weatherStr);
@@ -560,7 +564,7 @@ export default class OpenWeather extends Plugin {
     }
     if (this.settings.weatherFormat3.length > 0) {
       if (editor.contains("%weather3%")) {
-        let wstr = new FormatWeather (this.settings.location, this.settings.key, this.settings.units, this.settings.weatherFormat3);
+        let wstr = new FormatWeather (this.settings.location, this.settings.key, this.settings.units, this.settings.language, this.settings.weatherFormat3);
         let weatherStr = await wstr.getWeatherString();
         if (weatherStr.length == 0) {return};
         editor = editor.replace(/%weather3%/gmi, weatherStr);
@@ -569,7 +573,7 @@ export default class OpenWeather extends Plugin {
     }
     if (this.settings.weatherFormat4.length > 0) {
       if (editor.contains("%weather4%")) {
-        let wstr = new FormatWeather (this.settings.location, this.settings.key, this.settings.units, this.settings.weatherFormat4);
+        let wstr = new FormatWeather (this.settings.location, this.settings.key, this.settings.units, this.settings.language, this.settings.weatherFormat4);
         let weatherStr = await wstr.getWeatherString();
         if (weatherStr.length == 0) {return};
         editor = editor.replace(/%weather4%/gmi, weatherStr);
@@ -588,7 +592,7 @@ export default class OpenWeather extends Plugin {
           displayErrorMsg = false;
         }
       } else {
-        let wstr = new FormatWeather(this.settings.location, this.settings.key, this.settings.units, this.settings.weatherFormatSB);
+        let wstr = new FormatWeather(this.settings.location, this.settings.key, this.settings.units, this.settings.language, this.settings.weatherFormatSB);
         let weatherStr = await wstr.getWeatherString();
         if (weatherStr.length == 0) {return};
         this.statusBar.setText(weatherStr);
@@ -630,6 +634,7 @@ class InsertWeatherPicker extends SuggestModal<Commands> implements OpenWeatherS
   location: string;
   key: string;
   units: string;
+  language: string;
   excludeFolder: string;
   weatherFormat1: string;
   weatherFormat2: string;
@@ -642,11 +647,12 @@ class InsertWeatherPicker extends SuggestModal<Commands> implements OpenWeatherS
   command: string;
   format: string;
 
-  constructor(location: string, key: string, units: string, weatherFormat1: string, weatherFormat2: string, weatherFormat3: string, weatherFormat4: string) {
+  constructor(location: string, key: string, units: string, language: string, weatherFormat1: string, weatherFormat2: string, weatherFormat3: string, weatherFormat4: string) {
     super(app);
     this.location = location;
     this.key = key;
     this.units = units;
+    this.language = language;
     this.weatherFormat1 = weatherFormat1;
     this.weatherFormat2 = weatherFormat2;
     this.weatherFormat3 = weatherFormat3;
@@ -660,7 +666,7 @@ class InsertWeatherPicker extends SuggestModal<Commands> implements OpenWeatherS
       const md = view as MarkdownView;
       if (md.getMode() === 'source') {
         if (this.weatherFormat1.length > 0) {
-          let wstr = new FormatWeather (this.location, this.key, this.units, this.weatherFormat1);
+          let wstr = new FormatWeather (this.location, this.key, this.units, this.language, this.weatherFormat1);
           let weatherStr = await wstr.getWeatherString();
           if (weatherStr.length > 0) {
             this.weatherFormat1 = weatherStr;
@@ -673,7 +679,7 @@ class InsertWeatherPicker extends SuggestModal<Commands> implements OpenWeatherS
           }
         }
         if (this.weatherFormat2.length > 0) {
-          let wstr = new FormatWeather (this.location, this.key, this.units, this.weatherFormat2);
+          let wstr = new FormatWeather (this.location, this.key, this.units, this.language, this.weatherFormat2);
           let weatherStr = await wstr.getWeatherString();
           if (weatherStr.length > 0) {
             this.weatherFormat2 = weatherStr;
@@ -686,7 +692,7 @@ class InsertWeatherPicker extends SuggestModal<Commands> implements OpenWeatherS
           }
         }
         if (this.weatherFormat3.length > 0) {
-          let wstr = new FormatWeather (this.location, this.key, this.units, this.weatherFormat3);
+          let wstr = new FormatWeather (this.location, this.key, this.units, this.language, this.weatherFormat3);
           let weatherStr = await wstr.getWeatherString();
           if (weatherStr.length > 0) {
             this.weatherFormat3 = weatherStr;
@@ -699,7 +705,7 @@ class InsertWeatherPicker extends SuggestModal<Commands> implements OpenWeatherS
           }
         }
         if (this.weatherFormat4.length > 0) {
-          let wstr = new FormatWeather (this.location, this.key, this.units, this.weatherFormat4);
+          let wstr = new FormatWeather (this.location, this.key, this.units, this.language, this.weatherFormat4);
           let weatherStr = await wstr.getWeatherString();
           if (weatherStr.length > 0) {
             this.weatherFormat4 = weatherStr;
@@ -837,6 +843,64 @@ class OpenWeatherSettingsTab extends PluginSettingTab {
           await this.plugin.updateWeather();
         })
       .setValue(this.plugin.settings.units);
+      });
+
+    new Setting(containerEl)
+      .setName("Language")
+      .setDesc("Supported languages available")
+      .addDropdown(dropDown => {
+        dropDown.addOption('af', 'Afrikaans');
+        dropDown.addOption('al', 'Albanian');
+        dropDown.addOption('ar', 'Arabic');
+        dropDown.addOption('az', 'Azerbaijani');
+        dropDown.addOption('bg', 'Bulgarian');
+        dropDown.addOption('ca', 'Catalan');
+        dropDown.addOption('cz', 'Czech');
+        dropDown.addOption('da', 'Danish');
+        dropDown.addOption('de', 'German');
+        dropDown.addOption('el', 'Greek');
+        dropDown.addOption('en', 'English');
+        dropDown.addOption('eu', 'Basque');
+        dropDown.addOption('fa', 'Persian (Farsi)');
+        dropDown.addOption('fi', 'Finnish');
+        dropDown.addOption('fr', 'French');
+        dropDown.addOption('gl', 'Galician');
+        dropDown.addOption('he', 'Hebrew');
+        dropDown.addOption('hi', 'Hindi');
+        dropDown.addOption('hr', 'Croatian');
+        dropDown.addOption('hu', 'Hungarian');
+        dropDown.addOption('id', 'Indonesian');
+        dropDown.addOption('it', 'Italian');
+        dropDown.addOption('ja', 'Japanese');
+        dropDown.addOption('kr', 'Korean');
+        dropDown.addOption('la', 'Latvian');
+        dropDown.addOption('lt', 'Lithuanian');
+        dropDown.addOption('mk', 'Macedonian');
+        dropDown.addOption('no', 'Norwegian');
+        dropDown.addOption('nl', 'Dutch');
+        dropDown.addOption('pl', 'Polish');
+        dropDown.addOption('pt', 'Portuguese');
+        dropDown.addOption('pt_br', 'Português Brasil');
+        dropDown.addOption('ro', 'Romanian');
+        dropDown.addOption('ru', 'Russian');
+        dropDown.addOption('sv', 'Swedish');
+        dropDown.addOption('sk', 'Slovak');
+        dropDown.addOption('sl', 'Slovenian');
+        dropDown.addOption('sp', 'Spanish');
+        dropDown.addOption('sr', 'Serbian');
+        dropDown.addOption('th', 'Thai');
+        dropDown.addOption('tr', 'Turkish');
+        dropDown.addOption('ua', 'Ukrainian');
+        dropDown.addOption('vi', 'Vietnamese');
+        dropDown.addOption('zh_cn', 'Chinese Simplified');
+        dropDown.addOption('zh_tw', 'Chinese Traditional');
+        dropDown.addOption('zu', 'Zulu');
+        dropDown.onChange(async (value) => {
+          this.plugin.settings.language = value;
+          await this.plugin.saveSettings();
+          await this.plugin.updateWeather();
+        })
+      .setValue(this.plugin.settings.language);
       });
 
     // OpenWeatherSettingsTab - H2 Header - Exclude Folder (Exclude Folder for Template String Replacement)