+++ CHANGELOG.md
@@ -11,6 +11,56 @@
 
 <!-- ## [v-inc] ${YEAR4}-${MONTHNUMBER}-${DATE} -->
 
+## [1.4.0] 2022-03-18
+### Added
+- Language support for the following languages
+  - af - Afrikaans
+  - al - Albanian
+  - ar - Arabic
+  - az - Azerbaijani
+  - bg - Bulgarian
+  - ca - Catalan
+  - cz - Czech
+  - da - Danish
+  - de - German
+  - el - Greek
+  - en - English
+  - eu - Basque
+  - fa - Persian (Farsi)
+  - fi - Finnish
+  - fr - French
+  - gl - Galician
+  - he - Hebrew
+  - hi - Hindi
+  - hr - Croatian
+  - hu - Hungarian
+  - id - Indonesian
+  - it - Italian
+  - ja - Japanese
+  - kr - Korean
+  - la - Latvian
+  - lt - Lithuanian
+  - mk - Macedonian
+  - no - Norwegian
+  - nl - Dutch
+  - pl - Polish
+  - pt - Portuguese
+  - pt_br - Português Brasil
+  - ro - Romanian
+  - ru - Russian
+  - sv - Swedish
+  - sk - Slovak
+  - sl - Slovenian
+  - sp - Spanish
+  - sr - Serbian
+  - th - Thai
+  - tr - Turkish
+  - ua - uk Ukrainian
+  - vi - Vietnamese
+  - zh_cn - Chinese Simplified
+  - zh_tw - Chinese Traditional
+  - zu - Zulu
+
 ## [1.3.0] 2022-03-10
 ### Fixed
 - removed obsidian from ID